ci: :sparkles: Added balance inside table and link for challan no. in Report/order Summary
